refactor: Move files into their own packages
