+++ src/main/java/joe/task/TaskEvent.java
@@ -1,4 +1,4 @@
-package joe;
+package joe.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
