+++ src/main/java/joe/task/TaskTodo.java
@@ -1,4 +1,5 @@
-package joe;
+package joe.task;
+
 public class TaskTodo extends Task {
     public TaskTodo (String task) {
         super(task);
